Revert "ＳＥのスライダーの処理と、スライダー選択処理追加"
This reverts commit e9249d2691889b3407d9cc5ab55f0552dded4c94.
