--- ProjectTeamH/Assets/Haraguni/Scripts/SelectSlider.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f9b8b5cd6ba22054fb6a051c32a395a8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
