--- ProjectTeamH/Assets/Haraguni/Scripts/seUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 42bfd77f1376fc346ac203c2d1bb02fe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
