+++ ProjectTeamH/Assets/Haraguni/SoundUI.unity
@@ -112,58 +112,6 @@ NavMeshSettings:
     debug:
       m_Flags: 0
   m_NavMeshData: {fileID: 0}
---- !u!1 &23929699
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 23929700}
-  - component: {fileID: 23929701}
-  m_Layer: 5
-  m_Name: GameObject
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!224 &23929700
-RectTransform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 23929699}
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children:
-  - {fileID: 144456234}
-  - {fileID: 1339628663}
-  m_Father: {fileID: 1128572974}
-  m_RootOrder: 2
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0.5, y: 0.5}
-  m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -195.3, y: 97}
-  m_SizeDelta: {x: 100, y: 100}
-  m_Pivot: {x: 0.5, y: 0.5}
---- !u!114 &23929701
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 23929699}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f9b8b5cd6ba22054fb6a051c32a395a8, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  bgm_UI: {fileID: 647545414}
-  se_UI: {fileID: 616639265}
 --- !u!1 &57056459
 GameObject:
   m_ObjectHideFlags: 0
@@ -229,12 +177,12 @@ RectTransform:
   m_Children:
   - {fileID: 2004955239}
   - {fileID: 647545412}
-  m_Father: {fileID: 23929700}
-  m_RootOrder: 0
+  m_Father: {fileID: 1128572974}
+  m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 0, y: 0}
+  m_AnchoredPosition: {x: -195.3, y: 97}
   m_SizeDelta: {x: 100, y: 100}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!1 &182166668
@@ -427,7 +375,6 @@ GameObject:
   m_Component:
   - component: {fileID: 616639263}
   - component: {fileID: 616639264}
-  - component: {fileID: 616639265}
   m_Layer: 5
   m_Name: SE
   m_TagString: Untagged
@@ -503,31 +450,7 @@ MonoBehaviour:
   m_Value: 0
   m_OnValueChanged:
     m_PersistentCalls:
-      m_Calls:
-      - m_Target: {fileID: 616639265}
-        m_MethodName: MoveSlider
-        m_Mode: 1
-        m_Arguments:
-          m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
---- !u!114 &616639265
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 616639262}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 42bfd77f1376fc346ac203c2d1bb02fe, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  isSelect: 0
+      m_Calls: []
 --- !u!1 &647545411
 GameObject:
   m_ObjectHideFlags: 0
@@ -875,7 +798,8 @@ RectTransform:
   m_Children:
   - {fileID: 502658073}
   - {fileID: 1735068179}
-  - {fileID: 23929700}
+  - {fileID: 144456234}
+  - {fileID: 1339628663}
   m_Father: {fileID: 0}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -990,12 +914,12 @@ RectTransform:
   m_Children:
   - {fileID: 1224705662}
   - {fileID: 616639263}
-  m_Father: {fileID: 23929700}
-  m_RootOrder: 1
+  m_Father: {fileID: 1128572974}
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 0, y: -146}
+  m_AnchoredPosition: {x: -195.3, y: -49}
   m_SizeDelta: {x: 100, y: 100}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!1 &1359672469
