+++ ProjectTeamH/Assets/Shimada/prefabs/スイッチ式移動床.prefab.meta
